refactor: :recycle: add TableSheet and PivotSheet based on the SpreadSheet
