+++ packages/s2-core/__tests__/spreadsheet/merge-cells-spec.tsx
@@ -12,7 +12,7 @@ import {
   S2DataConfig,
   S2Options,
   SheetComponent,
-  SpreadSheet,
+  PivotSheet,
 } from '@/index';
 
 let data = getMockData('../data/tableau-supermarket.csv');
@@ -30,7 +30,7 @@ const getSpreadSheet = (
   dataCfg: S2DataConfig,
   options: S2Options,
 ) => {
-  return new SpreadSheet(dom, dataCfg, options);
+  return new PivotSheet(dom, dataCfg, options);
 };
 
 const baseDataCfg: S2DataConfig = {
