+++ packages/s2-core/__tests__/spreadsheet/pressure-test-spec.tsx
@@ -7,7 +7,7 @@ import {
   S2DataConfig,
   S2Options,
   SheetComponent,
-  SpreadSheet,
+  PivotSheet,
 } from '../../src';
 import { getContainer } from '../util/helpers';
 
@@ -29,7 +29,7 @@ const getSpreadSheet = (
   dataCfg: S2DataConfig,
   options: S2Options,
 ) => {
-  return new SpreadSheet(dom, dataCfg, options);
+  return new PivotSheet(dom, dataCfg, options);
 };
 
 const getDataCfg = () => {
