+++ packages/s2-core/__tests__/spreadsheet/table-sheet-spec.tsx
@@ -4,21 +4,22 @@ import 'antd/dist/antd.min.css';
 import React, { useEffect } from 'react';
 import ReactDOM from 'react-dom';
 import { act } from 'react-dom/test-utils';
+import { getContainer, getMockData } from '../util/helpers';
 import {
   S2DataConfig,
   S2Event,
   S2Options,
   SheetComponent,
   SpreadSheet,
-} from '../../src';
-import { getContainer, getMockData } from '../util/helpers';
+  PivotSheet,
+} from '@/index';
 
 const data = getMockData('../data/tableau-supermarket.csv');
 
 const getSpreadSheet =
   (ref) =>
   (dom: string | HTMLElement, dataCfg: S2DataConfig, options: S2Options) => {
-    const s2 = new SpreadSheet(dom, dataCfg, options);
+    const s2 = new PivotSheet(dom, dataCfg, options);
     ref.current = s2;
     return s2;
   };
