+++ packages/s2-core/__tests__/unit/data-accuracy-two-measures-spec.tsx
@@ -15,6 +15,7 @@ import { getContainer } from '../util/helpers';
 import {
   auto,
   EXTRA_FIELD,
+  PivotSheet,
   S2DataConfig,
   S2Options,
   SheetComponent,
@@ -29,7 +30,7 @@ const setSpreadSheet = (
   options: S2Options,
   index: number,
 ) => {
-  const ss = new SpreadSheet(dom, dataCfg, options);
+  const ss = new PivotSheet(dom, dataCfg, options);
   if (index === 1) {
     spreadsheet1 = ss;
   }
