+++ s2-site/examples/basic/pivot/demo/grid.ts
@@ -1,4 +1,4 @@
-import { SpreadSheet } from '@antv/s2';
+import { PivotSheet } from '@antv/s2';
 import '@antv/s2/dist/s2.min.css';
 
 fetch('../data/basic.json')
@@ -18,7 +18,7 @@ fetch('../data/basic.json')
       width: 600,
       height: 600,
     };
-    const s2 = new SpreadSheet(container, s2DataConfig, s2options);
+    const s2 = new PivotSheet(container, s2DataConfig, s2options);
 
     s2.render();
   });
