+++ s2-site/examples/basic/pivot/demo/tree.ts
@@ -1,4 +1,4 @@
-import { SpreadSheet } from '@antv/s2';
+import { PivotSheet } from '@antv/s2';
 import '@antv/s2/dist/s2.min.css';
 
 fetch('../data/basic.json')
@@ -19,7 +19,7 @@ fetch('../data/basic.json')
       height: 600,
       hierarchyType: 'tree',
     };
-    const s2 = new SpreadSheet(container, s2DataConfig, s2options);
+    const s2 = new PivotSheet(container, s2DataConfig, s2options);
 
     s2.render();
   });
